1. Rename BenchmarkParam and BenchmarkParams to ToolParam and ToolParams respectively, and branch them to the upper-level directory. 2. Also branch the benchmark/logging.h to the upper level directory.
Note we still keep those in benchmark directory for backward compatibility but by reusing the new definitions here. These utility classes are helpful to and going to be used in other tflite-related tools.

PiperOrigin-RevId: 305834516
