+++ tensorflow/lite/tools/benchmark/logging.h
@@ -16,74 +16,10 @@ limitations under the License.
 #ifndef TENSORFLOW_LITE_TOOLS_BENCHMARK_LOGGING_H_
 #define TENSORFLOW_LITE_TOOLS_BENCHMARK_LOGGING_H_
 
-// LOG and CHECK macros for benchmarks.
+// TODO(b/149482807): completely remove this file from the code base.
+#include "tensorflow/lite/tools/logging.h"
 
-#include <cstdlib>
-#include <iostream>
-#include <sstream>
-
-#ifdef _WIN32
-#undef ERROR
-#endif
-
-namespace tflite {
-namespace logging {
-// A wrapper that logs to stderr.
-//
-// Used for TFLITE_LOG and TFLITE_BENCHMARK_CHECK macros.
-class LoggingWrapper {
- public:
-  enum class LogSeverity : int {
-    INFO = 0,
-    WARN = 1,
-    ERROR = 2,
-    FATAL = 3,
-  };
-  LoggingWrapper(LogSeverity severity)
-      : severity_(severity), should_log_(true) {}
-  LoggingWrapper(LogSeverity severity, bool log)
-      : severity_(severity), should_log_(log) {}
-  std::stringstream& Stream() { return stream_; }
-  ~LoggingWrapper() {
-    if (should_log_) {
-      switch (severity_) {
-        case LogSeverity::INFO:
-        case LogSeverity::WARN:
-          std::cout << stream_.str() << std::endl;
-          break;
-        case LogSeverity::ERROR:
-          std::cerr << stream_.str() << std::endl;
-          break;
-        case LogSeverity::FATAL:
-          std::cerr << stream_.str() << std::endl;
-          std::flush(std::cerr);
-          std::abort();
-          break;
-      }
-    }
-  }
-
- private:
-  std::stringstream stream_;
-  LogSeverity severity_;
-  bool should_log_;
-};
-
-}  // namespace logging
-
-}  // namespace tflite
-
-#define TFLITE_LOG(severity)                                  \
-  tflite::logging::LoggingWrapper(                            \
-      tflite::logging::LoggingWrapper::LogSeverity::severity) \
-      .Stream()
-
-#define TFLITE_BENCHMARK_CHECK(condition)                  \
-  tflite::logging::LoggingWrapper(                         \
-      tflite::logging::LoggingWrapper::LogSeverity::FATAL, \
-      (condition) ? false : true)                          \
-      .Stream()
-
-#define TFLITE_BENCHMARK_CHECK_EQ(a, b) TFLITE_BENCHMARK_CHECK(a == b)
+#define TFLITE_BENCHMARK_CHECK(condition) TFLITE_TOOLS_CHECK(condition)
+#define TFLITE_BENCHMARK_CHECK_EQ(a, b) TFLITE_TOOLS_CHECK(a == b)
 
 #endif  // TENSORFLOW_LITE_TOOLS_BENCHMARK_LOGGING_H_
